requirements.txt
Django==3.2.6
PyJWT~=2.6.0
PyMySQL~=1.0.2
requests~=2.28.2